allow overwrite=False

+++ src/asim/extensions/visitor/visitor_tour_enum.py
@@ -3,7 +3,7 @@ import pandas as pd
 import numpy as np
 
 
-def create_tour_enumeration(tables, parameters):
+def create_tour_enumeration(tables, parameters, overwrite=True):
 
     # This is the main tour generation function.
     # 1. Generate n number of parties per segment
@@ -39,8 +39,9 @@ def create_tour_enumeration(tables, parameters):
     tours_enum = assign_person_households(tours)
 
     # Save output to corresponding csv
-    for name, output in tours_enum.items():
-        output.to_csv(os.path.join(parameters['config_dir'], parameters['output_fname'][name]))
+    if overwrite:
+        for name, output in tours_enum.items():
+            output.to_csv(os.path.join(parameters['config_dir'], parameters['output_fname'][name]))
 
     return tours_enum
